Refresh InstallerWindow
