+++ ToraConHelper/Installer/InstallerWindow.xaml
@@ -1,31 +1,58 @@
-﻿<Window
+﻿<ui:FluentWindow
     x:Class="ToraConHelper.Installer.InstallerWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:local="clr-namespace:ToraConHelper.Installer"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    WindowStyle="ToolWindow"
+    xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml"
     Title="ToraCon-Helper Telemetry DLL インストール"
     Width="800"
-    Height="450"
+    Height="600"
     FontSize="14"
     mc:Ignorable="d">
     <Window.Resources>
         <Style TargetType="Button">
-            <Setter Property="Margin" Value="5,0" />
+            <Setter Property="Margin" Value="5,5,0,0" />
             <Setter Property="Padding" Value="5"/>
         </Style>
     </Window.Resources>
     <Grid Margin="20">
         <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
             <RowDefinition Height="*" />
             <RowDefinition Height="Auto" />
         </Grid.RowDefinitions>
-        <ListBox ItemsSource="{Binding .}" Grid.Row="0" />
-        <StackPanel Orientation="Horizontal" HorizontalAlignment="Right" Grid.Row="1">
-            <Button Content="プラグインを手動でコピー" x:Name="pluginFolderButton" Click="pluginFolderButton_Click"/>
-            <Button Content="閉じる" x:Name="closeButton" Click="closeButton_Click"/>
+        <Grid.Resources>
+            <Style TargetType="ui:Button" BasedOn="{StaticResource {x:Type ui:Button}}">
+                <Setter Property="Margin" Value="5,0"/>
+            </Style>
+        </Grid.Resources>
+        <ui:TitleBar Title="{Binding RelativeSource={RelativeSource AncestorType=ui:FluentWindow}, Path=Title}" ShowMinimize="False" />
+        <TextBox
+            Grid.Row="1"
+            AcceptsReturn="True"
+            AcceptsTab="True"
+            IsReadOnly="True"
+            Margin="10,5"
+            Text="{Binding Message}" />
+        <StackPanel
+            Grid.Row="2"
+            HorizontalAlignment="Right"
+            Orientation="Horizontal"
+            Margin="0,10,10,10"
+            >
+            <ui:Button 
+                x:Name="pluginFolderButton"
+                Click="pluginFolderButton_Click"
+                Content="プラグインを手動でコピー"
+                Icon="{ui:SymbolIcon Symbol=WindowEdit20, Filled=True}"/>
+            <ui:Button
+                x:Name="closeButton"
+                Click="closeButton_Click"
+                Appearance="Primary"
+                Content="閉じる"
+                Icon="{ui:SymbolIcon Symbol=DismissSquare24, Filled=True}"/>
         </StackPanel>
     </Grid>
-</Window>
+</ui:FluentWindow>
