+++ ToraConHelper/Installer/InstallerWindow.xaml.cs
@@ -1,5 +1,5 @@
-﻿using System;
-using System.Collections.ObjectModel;
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+using System;
 using System.Diagnostics;
 using System.IO;
 using System.Threading.Tasks;
@@ -13,18 +13,24 @@ namespace ToraConHelper.Installer;
 /// </summary>
 public partial class InstallerWindow
 {
-    private readonly ObservableCollection<string> messages = [];
+    private readonly InstallerWindowViewModel viewModel = new();
     public InstallerWindow()
     {
         InitializeComponent();
 
-        DataContext = messages;
+        DataContext = viewModel;
 
         Loaded += InstallerWindow_Loaded;
     }
 
     private void InstallerWindow_Loaded(object sender, RoutedEventArgs e)
     {
+        Wpf.Ui.Appearance.SystemThemeWatcher.Watch(
+            this,                                    // Window class
+            Wpf.Ui.Controls.WindowBackdropType.Mica, // Background type
+            true                                     // Whether to change accents automatically
+        );
+
         _ = Task.Run(() =>
         {
             var pluginInstaller = new PluginInstaller();
@@ -37,7 +43,7 @@ public partial class InstallerWindow
     {
         this.Dispatcher.BeginInvoke(DispatcherPriority.Background, new Action(() =>
         {
-            messages.Add(e.Message);
+            viewModel.Message += $"{e.Message}{Environment.NewLine}";
         }));
     }
 
@@ -46,16 +52,9 @@ public partial class InstallerWindow
     private void pluginFolderButton_Click(object sender, RoutedEventArgs e)
         => Process.Start(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, Path.GetDirectoryName(PluginInstaller.SourcePath)));
 }
-public partial class PluginApp : Application
-{
-    public PluginApp() : base()
-    {
-        MainWindow = new InstallerWindow();
-    }
 
-    protected override void OnStartup(StartupEventArgs e)
-    {
-        base.OnStartup(e);
-        MainWindow.Show();
-    }
+public partial class InstallerWindowViewModel : ObservableObject
+{
+    [ObservableProperty]
+    private string message = string.Empty;
 }
