+++ ToraConHelperWap/Package.appxmanifest
@@ -15,7 +15,7 @@
   <Identity
     Name="1445kuramochia.ToraCon-Helper"
     Publisher="CN=33C05C0F-4C63-4E93-83D7-0AD7451F92EB"
-    Version="2.1.4.0" />
+    Version="2.1.5.0" />
 
   <Properties>
     <DisplayName>ToraCon-Helper</DisplayName>
